+++ example/lib/screens/forms.dart
@@ -218,6 +218,21 @@ class _FormsState extends State<Forms> {
           ),
         ),
         const SizedBox(height: 20),
+        InfoLabel(
+          label: 'Selectable Text',
+          child: Mica(
+            child: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: SelectableText(
+                'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
+                selectionControls: fluentTextSelectionControls,
+                showCursor: true,
+                cursorWidth: 1.5,
+              ),
+            ),
+          ),
+        ),
+        const SizedBox(height: 20),
       ],
     );
   }